backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1
python-multipart==0.0.6
psycopg2-binary==2.9.7
aiohttp==3.8.5
deepl==1.15.0
openai==0.27.8
google-generativeai==0.3.1